MAINT refactor ARFF parser (#22026)
